Refactor suite au déplacement dans le package services
